add queries about whether specific items are allowed by policies
For example, a bill to make a shield belt can be set for every colonist
that has an apparel policy allowing a shield belt. The advantage to
the existing policy queries is that this updates automatically whenever
policies are changed in any way.

One annoying detail of the drug query is that all drug policies by default
allow all drugs for addictions, even the no-drugs one, which is silly.
In order to the drug query to be usable this will need to be manually reset.
I have not come up with a better solution, having 3 queries for drugs
for the 3 allowed types seems an overkill, checking in-inventory is
unreliable as well.